snippets: import rng snippet update

+++ algo/snippets_code/random_snippet.cpp
@@ -2,6 +2,6 @@
 #define RANDOM_SNIPPET
 #include <random>
 #include <chrono>
-std::mt19937 rng(std::chrono::steady_clock::now().time_since_epoch().count());
+std::mt19937 rng((unsigned int)std::chrono::steady_clock::now().time_since_epoch().count());
 unsigned long long int grng(const unsigned long long int maxval){return rng()%maxval;}
 #endif
